La partie exception

+++ src/main/java/com/example/ebankback/services/BankAccountServiceImpl.java
@@ -3,6 +3,10 @@ package com.example.ebankback.services;
 import com.example.ebankback.dtos.*;
 import com.example.ebankback.entities.*;
 import com.example.ebankback.entities.ennum.OperationType;
+import com.example.ebankback.exceptions.BalanceNotSufficentException;
+import com.example.ebankback.exceptions.BankAccountNotFondEception;
+import com.example.ebankback.exceptions.CustomerNotFoundException;
+import com.example.ebankback.exceptions.IllegalArgumentException;
 import com.example.ebankback.mappers.BankAccountMapperImpl;
 import com.example.ebankback.repositories.AccountOperationRepository;
 import com.example.ebankback.repositories.BankAccountRepository;
@@ -34,8 +38,11 @@ public class BankAccountServiceImpl implements BankAccountSerivce{
     }
 
     @Override
-    public CurrentBankAccountDTO saveCurrentBankAccountDTO(double initialBalance, double overDraft, Long customerId) {
+    public CurrentBankAccountDTO saveCurrentBankAccountDTO(double initialBalance, double overDraft, Long customerId) throws CustomerNotFoundException {
         Customer customer = customerRepository.findById(customerId).orElse(null);
+        if(customer == null){
+              new CustomerNotFoundException("Customer not found");
+        }
         CurrentAccount currentAccount = new CurrentAccount();
         currentAccount.setId(UUID.randomUUID().toString());
         currentAccount.setCreatedAt(new Date());
@@ -47,8 +54,11 @@ public class BankAccountServiceImpl implements BankAccountSerivce{
     }
 
     @Override
-    public SavingBankAccountDTO saveSavingBankAccountDTO(double initialBalance, double interestRaye, Long customerId) {
+    public SavingBankAccountDTO saveSavingBankAccountDTO(double initialBalance, double interestRaye, Long customerId) throws CustomerNotFoundException {
         Customer customer = customerRepository.findById(customerId).orElse(null);
+        if(customer == null){
+            throw new CustomerNotFoundException("Customer not found");
+        }
         SavingAccount savingAccount = new SavingAccount();
         savingAccount.setId(UUID.randomUUID().toString());
         savingAccount.setCreatedAt(new Date());
@@ -75,9 +85,13 @@ public class BankAccountServiceImpl implements BankAccountSerivce{
     }
 
     @Override
-    public BankAccountDTO getBankAccount(String accountId) {
+    public BankAccountDTO getBankAccount(String accountId) throws BankAccountNotFondEception {
         BankAccount bankAccount = bankAccountRepository.findById(accountId)
-                .orElse(null);
+                .orElseThrow(()->new BankAccountNotFondEception("BankAccount not found"));
+
+        if(bankAccount == null){
+
+        }
         if(bankAccount instanceof SavingAccount){
             SavingAccount savingAccount = (SavingAccount) bankAccount;
             return dtoMapper.fromSavingBankAccount(savingAccount);
@@ -88,8 +102,11 @@ public class BankAccountServiceImpl implements BankAccountSerivce{
     }
 
     @Override
-    public void debit(String accountId, double amount, String description) {
+    public void debit(String accountId, double amount, String description) throws BalanceNotSufficentException {
         BankAccount bankAccount = bankAccountRepository.findById(accountId).orElse(null);
+        if(bankAccount.getBalance()<amount){
+            throw new BalanceNotSufficentException("Balance not sufficient");
+        }
         AccountOperation accountOperation = new AccountOperation();
         accountOperation.setType(OperationType.DEBIT);
         accountOperation.setAmount(amount);
@@ -101,8 +118,8 @@ public class BankAccountServiceImpl implements BankAccountSerivce{
     }
 
     @Override
-    public void credit(String accountId, double amount, String description) {
-        BankAccount bankAccount = bankAccountRepository.findById(accountId).orElse(null);
+    public void credit(String accountId, double amount, String description) throws BankAccountNotFondEception {
+        BankAccount bankAccount = bankAccountRepository.findById(accountId).orElseThrow(()->new BankAccountNotFondEception("Bank account not found"));
         AccountOperation accountOperation = new AccountOperation();
         accountOperation.setType(OperationType.CREDIT);
         accountOperation.setAmount(amount);
@@ -114,19 +131,24 @@ public class BankAccountServiceImpl implements BankAccountSerivce{
     }
 
     @Override
-    public void transfert(String accountIdSource, String accountIdDestination, double amount) {
+    public void transfert(String accountIdSource, String accountIdDestination, double amount) throws BalanceNotSufficentException, BankAccountNotFondEception {
         debit(accountIdSource, amount, "Transfert to "+accountIdDestination);
         credit(accountIdDestination,amount,"Treansfert from "+accountIdSource);
     }
 
     @Override
-    public List<BankAccountDTO> bankAccountList() {
-        return null;
+    public List<BankAccount> bankAccountList() {
+        return bankAccountRepository.findAll();
     }
 
     @Override
-    public CustomerDTO getCustomer(Long customerId) {
-        Customer customer = customerRepository.findById(customerId).orElse(null);
+    public CustomerDTO getCustomer(Long customerId) throws CustomerNotFoundException, IllegalArgumentException {
+
+        if(customerId<=0 || !String.valueOf(customerId).matches("\\d+") ){
+            throw new IllegalArgumentException("Invalid customer Id"+customerId);
+        }
+        Customer customer = customerRepository.findById(customerId)
+                .orElseThrow(()->new CustomerNotFoundException("Customer Not Found with id: "+ customerId));
         return dtoMapper.fromCustomer(customer);
     }
 
@@ -138,9 +160,14 @@ public class BankAccountServiceImpl implements BankAccountSerivce{
     }
 
     @Override
-    public CustomerDTO deleteCustomer(Long customerId) {
+    public CustomerDTO deleteCustomer(Long customerId) throws CustomerNotFoundException, IllegalArgumentException{
+        if(customerId<=0 || !String.valueOf(customerId).matches("\\d+") ){
+            throw new IllegalArgumentException("Invalid customer Id"+customerId);
+        }
+        Customer customer = customerRepository.findById(customerId)
+                .orElseThrow(()->new CustomerNotFoundException("Customer Not Found with id: "+ customerId));
         customerRepository.deleteById(customerId);
-        return null;
+        return dtoMapper.fromCustomer(customer);
     }
 
     @Override
